+++ src/config-selectors.js
@@ -49,6 +49,38 @@ mixitup.ConfigSelectors = function() {
 
     this.target = '.mix';
 
+    /**
+     * A optional selector string used to add further specificity to the querying of control elements,
+     * in addition to their mandatory data attribute (e.g. `data-filter`, `data-toggle`, `data-sort`).
+     *
+     * This can be used if other elements in your document must contain the above attributes
+     * (e.g. for use in third-party scripts), and would otherwise interfere with MixItUp. Adding
+     * an additional `control` selector of your choice allows MixItUp to restrict event handling
+     * to only those elements matching the defined selector.
+     *
+     * @name        control
+     * @memberof    mixitup.Config.selectors
+     * @instance
+     * @type        {string}
+     * @default     ''
+     *
+     * @example <caption>Example 1: Adding a `selectors.control` selector</caption>
+     *
+     * var mixer = mixitup(containerEl, {
+     *     selectors: {
+     *         control: '.mixitup-control'
+     *     }
+     * });
+     *
+     * // Will not be handled:
+     * // <button data-filter=".category-a"></button>
+     *
+     * // Will be handled:
+     * // <button class="mixitup-control" data-filter=".category-a"></button>
+     */
+
+    this.control = '';
+
     this.callActions('afterConstruct');
 
     h.seal(this);