+++ CHANGELOG.md
@@ -3,8 +3,14 @@ Change Log
 
 ## 3.0.0
 
-Release
+- Release
 
 ## 3.0.1
 
-Fixed issue where `layout.containerClassName` is not reflected in state object after instantiation.
+- Fixed issue where `layout.containerClassName` is not reflected in state object after instantiation.
+
+## 3.1.0
+
+- Added `selectors.controls` configuration option to allow for further specificity of control querying
+in addition to the mandatory data attirbutes.
+- Fixed package.json issues